+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.3.0 (2017-02-??)
 - Perfomance tweaks
 - Partial graph instead of complete one
+- FAQs, Docs and Help-Page are online
+- Search function over all statements
 
 ## 1.2.3 (2017-02-08)
 - Queue for duplicates